ai_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
SpeechRecognition==3.10.1
gTTS==2.5.1
pydub==0.25.1
ffmpeg-python==0.2.0
python-multipart==0.0.9
openai==1.12.0
python-dotenv==1.0.1
